GCZ: use 64-bit for m_data_offset
A malicious GCZ file could probably force this to be negative.
Shouldn't cause any issues other than file read failures, but need to fix
because it is causing errors on MSVC.

+++ Source/Core/DiscIO/CompressedBlob.h
@@ -72,7 +72,7 @@ private:
   CompressedBlobHeader m_header = {};
   std::vector<u64> m_block_pointers = {};
   std::vector<u32> m_hashes = {};
-  int m_data_offset = 0;
+  u64 m_data_offset = 0;
   File::DirectIOFile m_file = {};
   u64 m_file_size = 0;
   std::vector<u8> m_zlib_buffer = {};
